Ensure we always scale up if task data missing
Even if there are not enough healthy tasks to get a completely accurate
calculation on utilization we should still scale up if the remaining
tasks are overutilized. This should prevent the autoscaler becoming
"wedged" if there are a few unhealthy tasks lingering in a small pool of
total tasks. This should preserve the current behaviour if the
utilization suggests we need to scale down. That is we still bail and do
nothing rather than scaling down because it could be risky to do so with
incomplete data.