Make setup_marathon_job return bounce_again var
This adds an extra return variable to setup_marathon_job with a time in
which the service instance it is deploying should be bounced again. To
start with this aims to be simple and naive. If it thinks everything is
steady state it returns -1, if it thinks more work needs to be done it
returns 60 (seconds). This can be further nuanced in the future.

This extra return variable is consumed by the paasta deployd workers
which can then re-add services to the bounce queue. For the existing
cron jobs the extra variable is just ignored.